% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted_rand_index.R
\name{adjusted_rand_index}
\alias{adjusted_rand_index}
\title{Adjusted Rand Index between two partitions}
\usage{
adjusted_rand_index(part1, part2)
}
\arguments{
\item{part1, part2}{Two vectors of group labels (numeric, character, or
factor), of the same length, giving the partition of each object under
two different classifications.}
}
\value{
A single numeric value (ARI), equal to 1 for identical partitions
and close to 0 for partitions no more similar than chance.
}
\description{
Computes the Adjusted Rand Index (ARI) between two partitions of the same
set of objects, as defined by Hubert & Arabie (1985).
}
