% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ari_distance.R
\name{ari_distance}
\alias{ari_distance}
\title{Adjusted Rand Index distance}
\usage{
ari_distance(data, ind1, ind2)
}
\arguments{
\item{data}{A matrix or data.frame with products in rows and individuals
in columns (group labels), as used in \code{\link{total_dissim}}.}

\item{ind1}{the first individual}

\item{ind2}{the second individual}
}
\value{
The ARI distance between the two individuals
}
\description{
Adjusted Rand Index distance
}
