% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_ari.R
\name{mean_ari}
\alias{mean_ari}
\title{Mean Adjusted Rand Index between a partition and a set of partitions}
\usage{
mean_ari(partition, data)
}
\arguments{
\item{partition}{A vector of group labels for the candidate (consensus)
partition.}

\item{data}{A matrix or data.frame with products in rows and individuals
in columns (individual partitions / group labels).}
}
\value{
The mean ARI between \code{partition} and each column of
\code{data}.
}
\description{
Mean Adjusted Rand Index between a partition and a set of partitions
}
\keyword{internal}
