% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_index.R
\name{rand_index}
\alias{rand_index}
\title{Rand Index between two partitions}
\usage{
rand_index(part1, part2)
}
\arguments{
\item{part1, part2}{Two vectors of group labels (numeric, character, or
factor), of the same length, giving the partition of each object under
two different classifications.}
}
\value{
A single numeric value (RI), equal to 1 for identical partitions
and close to 0 for partitions no more similar than chance.
}
\description{
Computes the  Rand Index (RI) between two partitions of the same
set of objects.
}
